add prompt command

+++ index.js
@@ -1,5 +1,9 @@
 const admin = require('firebase-admin');
 const utils = require('./utils')
+const readline = require('readline').createInterface({
+    input: process.stdin,
+    output: process.stdout
+});
 /**
  * Make the backup (read collection & write to json, delete old backup, read json & save in backup collection)
  * @param db
@@ -7,7 +11,14 @@ const utils = require('./utils')
  * @param backupCollection
  */
 exports.backupCollection = function (db, myCollection, backupCollection) {
-    utils.backupCollectionToJson(db, myCollection).then(() => utils.deleteCollection(db, backupCollection).then(() => utils.saveBackupToFirestore(db, backupCollection)));
+    readline.question(`Are you sure you want to make a backup of ${myCollection} in the collection named ${backupCollection} ?[yes]/no : `, async function (answer) {
+        if (answer === 'no') {
+            console.log("Ok, bye.");
+            process.exit(1);
+        } else  {
+            utils.backupCollectionToJson(db, myCollection).then(() => utils.deleteCollection(db, backupCollection).then(() => utils.saveBackupToFirestore(db, backupCollection)));
+        }
+    });
 };
 
 /**
@@ -16,7 +27,14 @@ exports.backupCollection = function (db, myCollection, backupCollection) {
  * @param collection
  */
 exports.backupCollectionToJson = function (db, collection) {
-    utils.backupCollectionToJson(db, collection);
+    readline.question(`Are you sure you want to make a backup of ${myCollection} in backup.json ?[yes]/no : `, async function (answer) {
+        if (answer === 'no') {
+            console.log("Ok, bye.");
+            process.exit(1);
+        } else  {
+            await utils.backupCollectionToJson(db, collection);
+        }
+    });
 };
 
 /**
@@ -25,7 +43,14 @@ exports.backupCollectionToJson = function (db, collection) {
  * @param backupCollection
  */
 exports.deleteBackupCollection = function (db, backupCollection) {
-    utils.deleteCollection(db, backupCollection);
+    readline.question(`Are you sure you want to delete ${backupCollection} ?[yes]/no : `, async function (answer) {
+        if (answer === 'no') {
+            console.log("Ok, bye.");
+            process.exit(1);
+        } else  {
+            await utils.deleteCollection(db, backupCollection);
+        }
+    });
 };
 /**
  * copy json to database
@@ -33,7 +58,14 @@ exports.deleteBackupCollection = function (db, backupCollection) {
  * @param collection
  */
 exports.backupJsonToCollection = function (db, collection) {
-    utils.saveBackupToFirestore(db, collection);
+    readline.question(`Are you sure you want to copy backup.json in ${collection} ?[yes]/no : `, async function (answer) {
+        if (answer === 'no') {
+            console.log("Ok, bye.");
+            process.exit(1);
+        } else  {
+            await utils.saveBackupToFirestore(db, collection);
+        }
+    });
 }
 
 /**
@@ -45,6 +77,7 @@ exports.init = function (key) {
         admin.initializeApp({
             credential: admin.credential.cert(key)
         });
+        console.log('Successfully connected to Firebase !');
     } catch (e) {
         throw new Error(`It seems there's a problem : ${e.code}`);
     }
